修复通知过期的问题，实现跨机器端口映射变更

+++ internal/biz/queue_task.go
@@ -201,7 +201,7 @@ func (m *TaskUseCase) UpdateTask(ctx context.Context, task *Task) error {
 				return err
 			}
 			mapping.Status = int(task.Status)
-			_ = m.networkMappingRepo.UpdateNetworkMapping(ctx, mapping)
+			_ = m.networkMappingRepo.UpdateNetworkMapping(ctx, id, mapping)
 
 		case queue.TaskCmd_STORAGE_CREATE:
 			id, err := uuid.Parse(param.(*queue.StorageSetupTaskParamVO).Id)
@@ -284,14 +284,20 @@ func (m *TaskUseCase) UpdateTask(ctx context.Context, task *Task) error {
 				return err
 			}
 			mapping.Status = int(task.Status)
-			_ = m.networkMappingRepo.UpdateNetworkMapping(ctx, mapping)
+			_ = m.networkMappingRepo.UpdateNetworkMapping(ctx, id, mapping)
 
 		case queue.TaskCmd_NAT_PROXY_DELETE, queue.TaskCmd_NAT_VISITOR_DELETE:
 			id, err := uuid.Parse(param.(*queue.NatNetworkMappingTaskParamVO).Id)
 			if err != nil {
 				return err
 			}
-			err = m.networkMappingRepo.DeleteNetworkMapping(ctx, id)
+			mapping, err := m.networkMappingRepo.GetNetworkMapping(ctx, id)
+			if err != nil {
+				return err
+			}
+			if mapping.DeleteState == true {
+				err = m.networkMappingRepo.DeleteNetworkMapping(ctx, id)
+			}
 		case queue.TaskCmd_STORAGE_CREATE:
 			id, err := uuid.Parse(param.(*queue.StorageSetupTaskParamVO).Id)
 			if err != nil {
